setup.py: 'version' from `git describe --tags --long` instead of SVN revsion#
Previously (when the repository of this sofrware was on GoogleCode
and Subversion was used), in setup.py, the version number was taken
from its subversion database by using sqlite3.
Now that the repository was moved to GitHub, a change was made so that
the version number is deduced from Git tags (and therefore we need
so-called 'annotated tags').

Minor change: name = "mc_tools" -> "mc-tools" in setup.py.