The package structure was corrected according to http://python-packaging.readthedocs.org/
According to the document "How To Package Your Python Code",
  http://python-packaging.readthedocs.org/
the actual Python modules should be placed in a subdirectory
with the same name as the package.  The document also state
that hyphens should not be used in the Python package name.
Violation of the above rules causes the following two issues:
1. a module that contains hyphens in its name can not be
   imported e.g. by;
     import mc-tools       # invalid syntax
     import mc_tools       # No module named mc_tools
     import mc-tools.phits
2. a directory of a subpackage is placed just below the 'site-packages' directory
   (by installing `pip install git+https://github.com/kbat/mc-tools.git#egg=mc-tools`)
     site-packages/mcnp
   instead of
     site-packages/mc-tools/mcnp

To resolve the two issues, a directory named 'mc_tools' was
created and all the modules were moved below the directory.
The package name in 'setup.py' also was revert back from
'mc-tools' to 'mc_tools' (the package name listed by `pip list`
is still 'mc-tools'; the package can be uninstalled by
`pip uninstall mc_tools` as well as `pip uninstall mc-tools`).

With these fixes, the modules can be imported by e.g.
     import mc_tools.phits