First version of CINDER'90 converter added.
In this first version, only TABLE 4 (TOTAL ACTIVITY (CURIES) BY NUCLIDE) is supported, due to limited time for developing.
Few tests have been performed for the moment, so USE IT WITH CAUTION.

In order to properly read the information from the activity table, make sure that the input file for the activation script contains the following lines:

cinder_options
  russ 0

This option forces TABCODE to print a nuclide line also if all the activities are zero.
Further information about the russ options can be found in the CINDER manual.

	Nicolò Borghi.