A major update is contained in this commit.
Files have been rewritten to better match the coding conventions of ROOT.
In addition, a new structure has been given to the files, hoping to start the process towards the use of CMake for configuring and building.
A last modification in include/LinkDef.h should now support both compiling with ROOT5 and ROOT6.